Semi-complete / semi-broken port of songlist display to new content api
