+++ mps_youtube/content.py
@@ -1,6 +1,11 @@
 import math
+import copy
 
-from . import g, screen, c
+from . import g, screen, c, streams
+from .util import fmt_time, uea_pad, real_len
+from .config import Config
+
+import pafy
 
 # In the future, this could support more advanced features
 class Content:
@@ -45,6 +50,38 @@ class StringContent(LineContent):
         return count
 
 
+class SongList(LineContent):
+    def __init__(self, songs, length=None, msg=None, failmsg=None, loadmsg=None):
+        self._songs = songs
+        self._length = length
+        if length is None:
+            self._length = len(songs)
+        self._msg = msg
+        self._failmsg = failmsg
+        self._loadmsg = loadmsg
+
+    def get_text(self, s, e):
+        screen.update(content=logo(col=c.b),
+                message=(self._loadmsg or ''))
+
+        if callable(self._songs):
+            songs = self._songs(s, e)
+        else:
+            songs = self._songs[s:e]
+
+        if not songs:
+            g.message = failmsg or g.message
+            return
+
+        # preload first result url
+        streams.preload(songs[0], delay=0)
+
+        return _generate_songlist_display(songs)
+     
+    def get_count(self):
+        return self._length
+
+
 def page_msg(page=0):
     """ Format information about currently displayed page to a string. """
     if isinstance(g.content, PaginatedContent):
@@ -60,3 +97,120 @@ def page_msg(page=0):
                               page_count,
                               '>' if page + 1 < page_count else ']')
     return None
+
+
+def _get_user_columns():
+    """ Get columns from user config, return dict. """
+    total_size = 0
+    user_columns = Config.COLUMNS.get
+    user_columns = user_columns.replace(",", " ").split()
+
+    defaults = {"views": dict(name="viewCount", size=4, heading="View"),
+                "rating": dict(name="rating", size=4, heading="Rtng"),
+                "comments": dict(name="commentCount", size=4, heading="Comm"),
+                "date": dict(name="uploaded", size=8, heading="Date"),
+                "user": dict(name="uploaderName", size=10, heading="User"),
+                "likes": dict(name="likes", size=4, heading="Like"),
+                "dislikes": dict(name="dislikes", size=4, heading="Dslk"),
+                "category": dict(name="category", size=8, heading="Category")}
+
+    ret = []
+    for column in user_columns:
+        namesize = column.split(":")
+        name = namesize[0]
+
+        if name in defaults:
+            z = defaults[name]
+            nm, sz, hd = z['name'], z['size'], z['heading']
+
+            if len(namesize) == 2 and namesize[1].isdigit():
+                sz = int(namesize[1])
+
+            total_size += sz
+            cw = screen.getxy().width
+            if total_size < cw - 18:
+                ret.append(dict(name=nm, size=sz, heading=hd))
+
+    return ret
+
+
+def _generate_songlist_display(songs, song=False, zeromsg=None):
+    """ Generate list of choices from a song list."""
+    # pylint: disable=R0914
+    max_results = screen.getxy().max_results
+
+    if not songs:
+        g.message = zeromsg or "Enter /search-term to search or [h]elp"
+        return logo(c.g) + "\n\n"
+
+    have_meta = all(x.ytid in g.meta for x in songs)
+    user_columns = _get_user_columns() if have_meta else []
+    maxlength = max(x.length for x in songs)
+    lengthsize = 8 if maxlength > 35999 else 7
+    lengthsize = 5 if maxlength < 6000 else lengthsize
+    reserved = 9 + lengthsize + len(user_columns)
+    cw = screen.getxy().width
+    cw -= 1
+    title_size = cw - sum(1 + x['size'] for x in user_columns) - reserved
+    before = [{"name": "idx", "size": 3, "heading": "Num"},
+              {"name": "title", "size": title_size, "heading": "Title"}]
+    after = [{"name": "length", "size": lengthsize, "heading": "Time"}]
+    columns = before + user_columns + after
+
+    for n, column in enumerate(columns):
+        column['idx'] = n
+        column['sign'] = "-" if not column['name'] == "length" else ""
+
+    fmt = ["%{}{}s  ".format(x['sign'], x['size']) for x in columns]
+    fmtrow = fmt[0:1] + ["%s  "] + fmt[2:]
+    fmt, fmtrow = "".join(fmt).strip(), "".join(fmtrow).strip()
+    titles = tuple([x['heading'][:x['size']] for x in columns])
+    hrow = c.ul + fmt % titles + c.w
+    out = "\n" + hrow + "\n"
+
+    for n, x in enumerate(songs[:max_results]):
+        col = (c.r if n % 2 == 0 else c.p) if not song else c.b
+        details = {'title': x.title, "length": fmt_time(x.length)}
+        details = copy.copy(g.meta[x.ytid]) if have_meta else details
+        otitle = details['title']
+        details['idx'] = "%2d" % (n + 1)
+        details['title'] = uea_pad(columns[1]['size'], otitle)
+        cat = details.get('category') or '-'
+        details['category'] = pafy.get_categoryname(cat)
+        data = []
+
+        for z in columns:
+            fieldsize, field = z['size'], z['name']
+            if len(details[field]) > fieldsize:
+                details[field] = details[field][:fieldsize]
+
+            data.append(details[field])
+
+        line = fmtrow % tuple(data)
+        col = col if not song or song != songs[n] else c.p
+        line = col + line + c.w
+        out += line + "\n"
+
+    return out + "\n" * (5 - len(songs)) if not song else out
+
+
+def logo(col=None, version=""):
+    """ Return text logo. """
+    col = col if col else random.choice((c.g, c.r, c.y, c.b, c.p, c.w))
+    logo_txt = r"""                                             _         _
+ _ __ ___  _ __  ___       _   _  ___  _   _| |_ _   _| |__   ___
+| '_ ` _ \| '_ \/ __|_____| | | |/ _ \| | | | __| | | | '_ \ / _ \
+| | | | | | |_) \__ \_____| |_| | (_) | |_| | |_| |_| | |_) |  __/
+|_| |_| |_| .__/|___/      \__, |\___/ \__,_|\__|\__,_|_.__/ \___|
+          |_|              |___/"""
+    version = " v" + version if version else ""
+    logo_txt = col + logo_txt + c.w + version
+    lines = logo_txt.split("\n")
+    length = max(len(x) for x in lines)
+    x, y, _ = screen.getxy()
+    indent = (x - length - 1) // 2
+    newlines = (y - 12) // 2
+    indent, newlines = (0 if x < 0 else x for x in (indent, newlines))
+    lines = [" " * indent + l for l in lines]
+    logo_txt = "\n".join(lines) + "\n" * newlines
+    return "" if g.debug_mode else logo_txt
