+++ mps_youtube/util.py
@@ -8,6 +8,7 @@ import ctypes
 import subprocess
 import logging
 import time
+import unicodedata
 
 import pafy
 
@@ -189,3 +190,72 @@ def get_pafy(item, force=False, callback=None):
         dbg("%s%sgot new pafy object: %s%s" % (c.y, thread, p.title[:26], c.w))
         dbg("%s%sgot new pafy object: %s%s" % (c.y, thread, p.videoid, c.w))
         return p
+
+
+def fmt_time(seconds):
+    """ Format number of seconds to %H:%M:%S. """
+    hms = time.strftime('%H:%M:%S', time.gmtime(int(seconds)))
+    H, M, S = hms.split(":")
+
+    if H == "00":
+        hms = M + ":" + S
+
+    elif H == "01" and int(M) < 40:
+        hms = str(int(M) + 60) + ":" + S
+
+    elif H.startswith("0"):
+        hms = ":".join([H[1], M, S])
+
+    return hms
+
+
+def uea_pad(num, t, direction="<", notrunc=False):
+    """ Right pad with spaces taking into account East Asian width chars. """
+    direction = direction.strip() or "<"
+
+    t = ' '.join(t.split('\n'))
+
+    # TODO: Find better way of dealing with this?
+    if num <= 0:
+        return ''
+
+    if not notrunc:
+        # Truncate to max of num characters
+        t = t[:num]
+
+    if real_len(t) < num:
+        spaces = num - real_len(t)
+
+        if direction == "<":
+            t = t + (" " * spaces)
+
+        elif direction == ">":
+            t = (" " * spaces) + t
+
+        elif direction == "^":
+            right = False
+
+            while real_len(t) < num:
+                t = t + " " if right else " " + t
+                right = not right
+
+    return t
+
+
+def real_len(u, alt=False):
+    """ Try to determine width of strings displayed with monospace font. """
+    if not isinstance(u, str):
+        u = u.decode("utf8")
+
+    u = xenc(u) # Handle replacements of unsuported characters
+
+    ueaw = unicodedata.east_asian_width
+
+    if alt:
+        # widths = dict(W=2, F=2, A=1, N=0.75, H=0.5)  # original
+        widths = dict(N=.75, Na=1, W=2, F=2, A=1)
+
+    else:
+        widths = dict(W=2, F=2, A=1, N=1, H=0.5)
+
+    return int(round(sum(widths.get(ueaw(char), 1) for char in u)))
